feat: moved back handler to the hooks folder
