+++ plugins/login-plugin/src/Components/SignUp/SignUpTV/index.js
@@ -6,7 +6,7 @@ import { useDimensions } from "@applicaster/zapp-react-native-utils/reactHooks/l
 
 import ClientLogo from "../../UIComponents/ClientLogo";
 import { mapKeyToStyle } from "../../../Utils/Customization";
-import { useBackHandler } from "../../../Utils/hooks";
+import { useBackHandler } from "../../../Utils/Hooks";
 import SignupControls from "./SignupControls";
 import { validateSignUpData } from "../../../Utils/Account";
