+++ plugins/login-plugin/src/Components/UIComponents/PrivacyPolicy/PrivacyPolicyTv/index.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import * as R from "ramda";
 import PrivacyTitle from "./PrivacyTitle";
 import PrivacyDescription from "./PrivacyDescription";
-import { useBackHandler } from "../../../../Utils/hooks";
+import { useBackHandler } from "../../../../Utils/Hooks";
 import { TVEventHandlerComponent } from "@applicaster/zapp-react-native-tvos-ui-components/Components/TVEventHandlerComponent";
 
 const PrivacyPolicyTv = (props) => {
