+++ plugins/login-plugin/src/Utils/Hooks/index.js
@@ -1,0 +1,1 @@
+export { useBackHandler } from "./useBackHandler";
